doc: active text truncate w/ dark mode
